+++ tracing_node.cpp
@@ -28,7 +28,7 @@ void TracingNode::init_traces(size_t num_traces)
     traces_.resize(num_traces);
     const size_t trace_length = ConnectionUpdateManager::instance()->get_max_latency();
     
-    for (int i=0; i<num_traces; i++)
+    for (size_t i=0; i<num_traces; i++)
     {
         traces_[i].resize(trace_length, 0.0);
     }
